Update README.md
Welcome to the South Korea Population Dynamics Analysis repository! This project provides an in-depth analysis of the population dynamics in South Korea from 2004 to 2023 using the logistic population model. By estimating key parameters such as the growth rate and carrying capacity, we aim to predict South Korea's population for the years 2030, 2040, and 2050.